Added a combat-check to the queue since most godvoice commands are useless in combat. If you want to send a command in combat, you'll need to do so manually.
Updated the README to reflect the combat-check changes for the queue.

Increased the maximum amount of data that can be collected and improved the culling procedure for when the data reaches this maximum. This should fix the issue where the data suddenly becomes extremely inaccurate after extended use. If not, then adjustments will be made.